+++ src/builtins/history.rs
@@ -2,7 +2,7 @@
 
 use crate::builtins::error::Error;
 use crate::history::in_private_mode;
-use crate::history::{self, History, history_session_id};
+use crate::history::{self, History, history_id};
 use crate::reader::commandline_get_state;
 use crate::{err_fmt, err_str};
 
@@ -256,7 +256,7 @@ pub fn history(parser: &Parser, streams: &mut IoStreams, args: &mut [&wstr]) -> 
     // from webconfig.py.
     let history = commandline_get_state(true)
         .history
-        .unwrap_or_else(|| History::new(&history_session_id(parser.vars())));
+        .unwrap_or_else(|| History::new(history_id(parser.vars())));
 
     // If a history command hasn't already been specified via a flag check the first word.
     // Note that this can be simplified after we eliminate allowing subcommands as flags.