+++ src/complete.rs
@@ -12,7 +12,7 @@ use crate::{
     },
     flog::{flog, flogf},
     function,
-    history::{History, history_session_id},
+    history::{History, history_id},
     localization::{LocalizableString, localizable_string},
     operation_context::OperationContext,
     parse_constants::SourceRange,
@@ -1675,7 +1675,7 @@ impl<'ctx> Completer<'ctx> {
                 // $history can be huge, don't put all of it in the completion description; see
                 // #6288.
                 if env_name == "history" {
-                    let history = History::new(&history_session_id(self.ctx.vars()));
+                    let history = History::new(history_id(self.ctx.vars()));
                     for i in 1..std::cmp::min(history.size(), 64) {
                         if i > 1 {
                             desc.push(' ');