+++ src/env_dispatch.rs
@@ -231,8 +231,8 @@ fn handle_term_size_change(vars: &EnvStack) {
 }
 
 fn handle_fish_history_change(vars: &EnvStack) {
-    let session_id = crate::history::history_session_id(vars);
-    reader_change_history(&session_id);
+    let history_id = crate::history::history_id(vars);
+    reader_change_history(history_id);
 }
 
 fn handle_fish_cursor_selection_mode_change(vars: &EnvStack) {